add a separate module for nav bar and import it in index.js file
