+++ src/style.css
@@ -2,7 +2,9 @@ body {
   background-image: url(./images/resturant-back.jpg);
   background-repeat: no-repeat;
   background-position: center;
+  margin: 0;
 }
+
 nav .nav-bar {
   display: flex;
   justify-content: space-between;
@@ -11,6 +13,7 @@ nav .nav-bar {
   font-weight: bolder;
   background-color: black;
   border-radius: 10px;
+  margin: 0;
 }
 
 .nav-bar a {
